Abstract class for a scanner fragment
