+++ app/src/main/res/layout/activity_main.xml
@@ -20,7 +20,7 @@
 
     <fragment
         android:id="@+id/scan_fragment"
-        class="shakram02.ahmed.prola.ScanFragment"
+        class="shakram02.ahmed.prola.CameraScanFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         android:layout_marginBottom="8dp"
@@ -31,7 +31,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/relativeLayout"
-        tools:layout="@layout/fragment_capture">
+        tools:layout="@layout/fragment_camera_scan">
 
     </fragment>
