+++ app/src/main/res/layout/fragment_camera_scan.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="shakram02.ahmed.prola.ScanFragment">
+    tools:context="shakram02.ahmed.prola.CameraScanFragment">
 
     <LinearLayout
         android:layout_width="0dp"
